apps/vision-lab/requirements.txt
fastapi>=0.116,<1.0
opencv-python-headless>=4.10,<5.0
pillow>=11.3,<12.0
pydantic>=2.11,<3.0
streamlit>=1.45,<2.0
ultralytics>=8.3,<9.0
uvicorn>=0.35,<1.0
